2022/day09/day09_test.go
package day09_test

import (
	_ "embed"
	"strings"
	"testing"

	day "github.com/minyiky/advent-of-code/2022/day09"
	"github.com/stretchr/testify/assert"
)

//go:embed input1_test.txt
var input1 string

//go:embed input2_test.txt
var input2 string

func SetUp(input string) []string {
	input = strings.ReplaceAll(input, "\r", "")
	lines := strings.Split(input, "\n")
	return lines
}

func Test_Part1(t *testing.T) {
	lines := SetUp(input1)
	expected := 13

	val, err := day.Part1Val(lines)

	assert.NoError(t, err)
	assert.Equal(t, expected, val)
}

func Test_Part2(t *testing.T) {
	lines := SetUp(input2)
	expected := 36

	val, err := day.Part2Val(lines)

	assert.NoError(t, err)
	assert.Equal(t, expected, val)
}
